work in progress

+++ src/main/java/com/inditex/test/product/domain/events/PriceChangedEvent.java
@@ -6,18 +6,19 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Getter @NoArgsConstructor @ToString(callSuper = true)
 public class PriceChangedEvent extends DomainEvent implements Serializable
 {
-    private float oldPrice;
-    private float newPrice;
+    private BigDecimal oldPrice;
+    private BigDecimal newPrice;
 
     // BUILDER:
     //--------------------------------------------------------------------------------------------------------
 
     @Builder
-    public PriceChangedEvent(Long aggregateId, float oldPrice, float newPrice)
+    public PriceChangedEvent(Long aggregateId, BigDecimal oldPrice, BigDecimal newPrice)
     {
         super(aggregateId, PriceChangedEvent.class);
         this.oldPrice = oldPrice;
